some modiffication

+++ DTOs/TokenDTO.cs
@@ -49,7 +49,7 @@
         public string DdoCode {get; set; }
         public string PayeeDept { get; set; }
         public HOAChain HOAChain { get; set; }
-        public double? GrossAmount { get; set; }
-        public double? NetAmount { get; set; }
+        public decimal? GrossAmount { get; set; }
+        public decimal? NetAmount { get; set; }
     }
 }
